Coverity ID 333614 sign extension
An intermediate calculation results in a cast to 'int' which is then
promoted (on 64-bit systems) to an unsigned long long. If the top bit
of the intermediate value is set, then all the top bits of the final
value will be set.

Do an explicit cast to an unsigned to avoid this.

+++ devices/gdevxalt.c
@@ -253,7 +253,7 @@ x_wrap_copy_color(gx_device * dev, const byte * base, int sourcex,
                         for (k = 0; k < source_bytes; k++) {
                             byte color = temp_color1 & 0xff;
                             temp_color1 >>= 8;
-                            temp_color2 += color << ((source_bytes - k - 1) * 8);
+                            temp_color2 += (uint32_t)(color << ((source_bytes - k - 1) * 8));
                         }
                         cindex = x_alt_map_color(dev, temp_color2);
                     }
